+++ src/components/ChannelInner.js
@@ -13,7 +13,6 @@ import { emojiData } from '../utils';
 import { LoadingIndicator } from './LoadingIndicator';
 import { LoadingErrorIndicator } from './LoadingErrorIndicator';
 import { EmptyStateIndicator } from './EmptyStateIndicator';
-import { KeyboardCompatibleView } from './KeyboardCompatibleView';
 import { logChatPromiseExecution } from 'stream-chat';
 
 /**
@@ -124,6 +123,10 @@ export class ChannelInner extends PureComponent {
     isOnline: PropTypes.bool,
     Message: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
     Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
+    /** Override send message request (Advanced usage only) */
+    doSendMessageRequest: PropTypes.func,
+    /** Override update message request (Advanced usage only) */
+    doUpdateMessageRequest: PropTypes.func,
   };
 
   static defaultProps = {
@@ -352,6 +355,20 @@ export class ChannelInner extends PureComponent {
     return message;
   };
 
+  // eslint-disable-next-line require-await
+  editMessage = async (updatedMessage) => {
+    if (this.props.doUpdateMessageRequest) {
+      return Promise.resolve(
+        this.props.doUpdateMessageRequest(
+          this.props.channel.cid,
+          updatedMessage,
+        ),
+      );
+    }
+
+    return this.props.client.updateMessage(updatedMessage);
+  };
+
   _sendMessage = async (message) => {
     // Scrape the reserved fields if present.
     const {
@@ -380,7 +397,16 @@ export class ChannelInner extends PureComponent {
     };
 
     try {
-      const messageResponse = await this.props.channel.sendMessage(messageData);
+      let messageResponse;
+      if (this.props.doSendMessageRequest) {
+        messageResponse = await this.props.doSendMessageRequest(
+          this.props.channel.cid,
+          messageData,
+        );
+      } else {
+        messageResponse = await this.props.channel.sendMessage(messageData);
+      }
+
       // replace it after send is completed
       if (messageResponse.message) {
         messageResponse.message.status = 'received';
@@ -566,6 +592,7 @@ export class ChannelInner extends PureComponent {
     updateMessage: this.updateMessage,
     removeMessage: this.removeMessage,
     sendMessage: this.sendMessage,
+    editMessage: this.editMessage,
     retrySendMessage: this.retrySendMessage,
     setEditingState: this.setEditingState,
     clearEditingState: this.clearEditingState,
@@ -593,7 +620,7 @@ export class ChannelInner extends PureComponent {
 
   render() {
     let core;
-
+    const { KeyboardCompatibleView } = this.props;
     if (this.state.error) {
       this.props.logger(
         'Channel component',
@@ -617,7 +644,9 @@ export class ChannelInner extends PureComponent {
       );
     } else {
       core = (
-        <KeyboardCompatibleView>
+        <KeyboardCompatibleView
+          enabled={!this.props.disableKeyboardCompatibleView}
+        >
           <ChannelContext.Provider value={this.getContext()}>
             <SuggestionsProvider
               handleKeyboardAvoidingViewEnabled={(trueOrFalse) => {