+++ src/components/MessageInput.js
@@ -18,11 +18,11 @@ import uniq from 'lodash/uniq';
 import styled from '@stream-io/styled-components';
 import { themed } from '../styles/theme';
 import { SendButton } from './SendButton';
+import { AttachButton } from './AttachButton';
 
 import { ActionSheetCustom as ActionSheet } from 'react-native-actionsheet';
 // import iconMedia from '../images/icons/icon_attach-media.png';
 
-import iconAddAttachment from '../images/icons/plus-outline.png';
 import iconGallery from '../images/icons/icon_attach-media.png';
 import iconFolder from '../images/icons/icon_folder.png';
 import iconClose from '../images/icons/icon_close.png';
@@ -59,17 +59,6 @@ const InputBoxContainer = styled.View`
   margin: 10px;
   align-items: center;
   ${({ theme }) => theme.messageInput.inputBoxContainer.css}
-`;
-
-const AttachButton = styled.TouchableOpacity`
-  margin-right: 8;
-  ${({ theme }) => theme.messageInput.attachButton.css}
-`;
-
-const AttachButtonIcon = styled.Image`
-  width: 15;
-  height: 15;
-  ${({ theme }) => theme.messageInput.attachButtonIcon.css}
 `;
 
 const ActionSheetTitleContainer = styled.View`
@@ -185,6 +174,15 @@ const MessageInput = withKeyboardContext(
               PropTypes.elementType,
             ]),
             /**
+             * Custom UI component for attach button.
+             *
+             * Defaults to and accepts same props as: [AttachButton](https://getstream.github.io/stream-chat-react-native/#attachbutton)
+             * */
+            AttachButton: PropTypes.oneOfType([
+              PropTypes.node,
+              PropTypes.elementType,
+            ]),
+            /**
              * Additional props for underlying TextInput component. These props will be forwarded as it is to TextInput component.
              *
              * @see See https://facebook.github.io/react-native/docs/textinput#reference
@@ -209,6 +207,7 @@ const MessageInput = withKeyboardContext(
             hasImagePicker: true,
             hasFilePicker: true,
             SendButton,
+            AttachButton,
           };
 
           getMessageDetailsForState = (message) => {
@@ -378,11 +377,9 @@ const MessageInput = withKeyboardContext(
               // TODO: Remove this line and show an error when submit fails
               this.props.clearEditingState();
 
-              const updateMessagePromise = this.props.client
-                .updateMessage(updatedMessage)
-                .then(() => {
-                  this.props.clearEditingState();
-                });
+              const updateMessagePromise = this.props
+                .editMessage(updatedMessage)
+                .then(this.props.clearEditingState);
               logChatPromiseExecution(updateMessagePromise, 'update message');
             } else {
               try {
@@ -408,7 +405,7 @@ const MessageInput = withKeyboardContext(
 
           updateMessage = async () => {
             try {
-              await this.props.client.updateMessage({
+              await this.props.client.editMessage({
                 ...this.props.editing,
                 text: this.state.text,
               });
@@ -696,7 +693,12 @@ const MessageInput = withKeyboardContext(
             this.attachActionSheet.hide();
           };
           render() {
-            const { hasImagePicker, hasFilePicker, SendButton } = this.props;
+            const {
+              hasImagePicker,
+              hasFilePicker,
+              SendButton,
+              AttachButton,
+            } = this.props;
             let editingBoxStyles = {};
             if (this.props.editing) {
               editingBoxStyles = {
@@ -755,7 +757,7 @@ const MessageInput = withKeyboardContext(
                     )}
                     <InputBoxContainer ref={this.props.setInputBoxContainerRef}>
                       <AttachButton
-                        onPress={async () => {
+                        handleOnPress={async () => {
                           if (hasImagePicker && hasFilePicker) {
                             await this.props.dismissKeyboard();
                             this.attachActionSheet.show();
@@ -764,9 +766,7 @@ const MessageInput = withKeyboardContext(
                           else if (!hasImagePicker && hasFilePicker)
                             this._pickFile();
                         }}
-                      >
-                        <AttachButtonIcon source={iconAddAttachment} />
-                      </AttachButton>
+                      />
                       {/**
                     TODO: Use custom action sheet to show icon with titles of button. But it doesn't
                     work well with async onPress operations. So find a solution.
@@ -836,7 +836,7 @@ const MessageInput = withKeyboardContext(
                         }
                       />
                       <SendButton
-                        title="Pick an image from camera roll"
+                        title="Send message"
                         sendMessage={this.sendMessage}
                         editing={this.props.editing}
                       />